requirements.txt
emoji
pandas
Pillow
python-telegram-bot
XlsxWriter
SQLAlchemy
psutil
python-telegram-bot-calendar
holidays
openpyxl